requirements.txt
backports.ssl-match-hostname==3.4.0.2
certifi==14.05.14
pyanybar==0.1.1
requests==2.6.0
six==1.9.0
tornado==4.1
websocket-client==0.29.0
wsgiref==0.1.2
